move  macro to same file as
